requirements.txt
asgiref==3.8.1
autopep8==2.1.0
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.7
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
psycopg==3.1.18
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
